# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Мониторинг интернета ВКО",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/map/schools": {
      "get": {
        "tags": [
          "map"
        ],
        "summary": "Школы на карте ВКО: GeoJSON с фильтрами",
        "description": "GeoJSON FeatureCollection (RFC 7946) активных школ в области видимости, без пагинации. Фильтры provider_id и connection_type_id отбирают школы, у которых есть такая линия. Статус — на момент period_to (ADR-004), как в сводке; показатели — последний замер основной линии в периоде, без Wi-Fi (ADR-012). Школа без координат приходит с geometry = null.",
        "operationId": "get_school_map",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "region_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Id"
            }
          },
          {
            "name": "provider_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Id"
            }
          },
          {
            "name": "connection_type_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Connection Type Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchoolStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Статусы школы; несколько — повтором параметра",
              "title": "Status"
            },
            "description": "Статусы школы; несколько — повтором параметра"
          },
          {
            "name": "period_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Начало периода, включительно; по умолчанию — без нижней границы",
              "title": "Period From"
            },
            "description": "Начало периода, включительно; по умолчанию — без нижней границы"
          },
          {
            "name": "period_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Конец периода, не включая; по умолчанию — текущий момент",
              "title": "Period To"
            },
            "description": "Конец периода, не включая; по умолчанию — текущий момент"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchoolMapFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/schools": {
      "get": {
        "tags": [
          "schools"
        ],
        "summary": "Список школ со средними показателями и статусом",
        "description": "Фильтры provider_id и connection_type_id отбирают школы, у которых есть такая линия. Сортировка серверная; без sort — по названию.",
        "operationId": "list_schools",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "region_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Id"
            }
          },
          {
            "name": "provider_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Id"
            }
          },
          {
            "name": "connection_type_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Connection Type Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchoolStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "true — активные, false — отключённые; не задан — все",
              "title": "Is Active"
            },
            "description": "true — активные, false — отключённые; не задан — все"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "description": "Название или School ID",
              "title": "Q"
            },
            "description": "Название или School ID"
          },
          {
            "name": "period_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Начало окна средних, включительно; по умолчанию — последние 24 ч, как на главной",
              "title": "Period From"
            },
            "description": "Начало окна средних, включительно; по умолчанию — последние 24 ч, как на главной"
          },
          {
            "name": "period_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Конец окна средних, не включая; по умолчанию — текущий момент",
              "title": "Period To"
            },
            "description": "Конец окна средних, не включая; по умолчанию — текущий момент"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SchoolSort"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Поле; «-» — по убыванию. status: normal < unstable < critical < offline, no_data в конце; -status — сначала худшие",
              "title": "Sort"
            },
            "description": "Поле; «-» — по убыванию. status: normal < unstable < critical < offline, no_data в конце; -status — сначала худшие"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Номер страницы, с 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Номер страницы, с 1"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Размер страницы"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchoolListItemPage"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "schools"
        ],
        "summary": "Создать школу",
        "description": "Неизвестный region_id — 422.",
        "operationId": "create_school",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchoolCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchoolDetail"
                }
              }
            }
          },
          "409": {
            "description": "School ID уже занят другой школой (type school_code_taken)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/schools/{school_id}": {
      "get": {
        "tags": [
          "schools"
        ],
        "summary": "Карточка школы: статус и текущие показатели",
        "operationId": "get_school",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "school_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "School Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchoolDetail"
                }
              }
            }
          },
          "404": {
            "description": "Школа не найдена",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "schools"
        ],
        "summary": "Изменить или деактивировать школу",
        "description": "Рабочие часы (working_hours) настраиваются в T-37. Неизвестный region_id — 422.",
        "operationId": "update_school",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "school_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "School Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchoolUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchoolDetail"
                }
              }
            }
          },
          "404": {
            "description": "Школа не найдена",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "School ID уже занят другой школой (type school_code_taken)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/schools/{school_id}/devices": {
      "get": {
        "tags": [
          "schools"
        ],
        "summary": "Компьютеры школы с последними замерами",
        "operationId": "list_school_devices",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "school_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "School Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Номер страницы, с 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Номер страницы, с 1"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Размер страницы"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceListItemPage"
                }
              }
            }
          },
          "404": {
            "description": "Школа не найдена",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/schools/{school_id}/lines": {
      "get": {
        "tags": [
          "schools"
        ],
        "summary": "Линии школы с договорными значениями",
        "operationId": "list_school_lines",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "school_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "School Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Номер страницы, с 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Номер страницы, с 1"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Размер страницы"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineDetailPage"
                }
              }
            }
          },
          "404": {
            "description": "Школа не найдена",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "schools"
        ],
        "summary": "Добавить линию школы",
        "description": "Неизвестный provider_id или connection_type_id — 422.",
        "operationId": "create_school_line",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "school_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "School Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineDetail"
                }
              }
            }
          },
          "404": {
            "description": "Школа не найдена",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/schools/{school_id}/lines/{line_id}": {
      "patch": {
        "tags": [
          "schools"
        ],
        "summary": "Изменить линию школы",
        "description": "Неизвестный provider_id или connection_type_id — 422.",
        "operationId": "update_school_line",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "school_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "School Id"
            }
          },
          {
            "name": "line_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Line Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineDetail"
                }
              }
            }
          },
          "404": {
            "description": "Школа или линия не найдены",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SchoolMapFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "const": "FeatureCollection",
            "title": "Type"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/SchoolMapFeature"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "type",
          "features"
        ],
        "title": "SchoolMapFeatureCollection",
        "description": "Every active school matching the filters: the map is a whole set, not a paginated list."
      },
...
      "SchoolStatus": {
        "type": "string",
        "enum": [
          "normal",
          "unstable",
          "critical",
          "offline",
          "no_data"
        ]
      },
...
      "ValidationProblem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Стабильный машинный код ошибки",
            "examples": [
              "validation_error"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Название ошибки для алерта",
            "examples": [
              "Ошибка валидации"
            ]
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "examples": [
              422
            ]
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Текст для человека"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "Путь запроса"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "errors"
        ],
        "title": "ValidationProblem",
        "description": "Request validation error (HTTP 422, ``type = validation_error``)."
      }
    },
    "securitySchemes": {
      "DeviceToken": {
        "type": "apiKey",
        "description": "Токен устройства: `Authorization: Device <token>` (ADR-005)",
        "in": "header",
        "name": "Authorization"
      },
      "RefreshCookie": {
        "type": "apiKey",
        "description": "Refresh-токен панели в httpOnly cookie",
        "in": "cookie",
        "name": "refresh_token"
      },
      "BearerAuth": {
        "type": "http",
        "description": "Access-токен панели: `Authorization: Bearer <jwt>`, живёт 15 минут",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}